REFACTOR: Move equality into model folder
